bookingsite/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.0
black==23.12.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.5
defusedxml==0.8.0rc2
distlib==0.3.8
Django==4.2.5
django-appconf==1.0.6
django-compressor==4.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-jazzmin==2.6.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
dnspython==2.4.2
drf-spectacular==0.27.0
frozenlist==1.4.1
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==5.3.4
Markdown==3.5.1
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pymongo==4.6.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
rcssmin==1.1.1
redis==5.0.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
rpds-py==0.13.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
yarl==1.9.4
urllib3==2.0.3
vine==5.0.0
virtualenv==20.24.5
wcwidth==0.2.8
xlrd==2.0.1
xlwt==1.3.0
django-jazzmin==2.6.0

